requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu121
torch==2.1.1+cu121
pillow==10.0.1
pyqt6==6.6.1
pyqt6-webengine==6.6.0
pyqt6-3d==6.6.0
qtpy==2.4.1
requests==2.31.0
pyqtgraph==0.13.3
pyyaml==6.0.1
https://github.com/chengzeyi/stable-fast/releases/download/v1.0.1/stable_fast-1.0.1+torch211cu121-cp310-cp310-win_amd64.whl
diffusers==0.25.0
transformers==4.36.0
xformers==0.0.23
omegaconf
opencv-contrib-python
pandas
numexpr
timm==0.6.5
basicsr
matplotlib
rich
psutil
PyQt6-QScintilla
pyflakes
aiohttp
torchsde
einops
openai
librosa
audiocraft
accelerate